+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/cases/entity/CaseBuilding.java
@@ -52,7 +52,7 @@ public class CaseBuilding {
 
     private BigDecimal diameterDepth;
 
-    private Integer landUseYear;
+    private BigDecimal landUseYear;
 
     private BigDecimal netHeight;
 
@@ -300,11 +300,11 @@ public class CaseBuilding {
         this.diameterDepth = diameterDepth;
     }
 
-    public Integer getLandUseYear() {
+    public BigDecimal getLandUseYear() {
         return landUseYear;
     }
 
-    public void setLandUseYear(Integer landUseYear) {
+    public void setLandUseYear(BigDecimal landUseYear) {
         this.landUseYear = landUseYear;
     }
